backend/requirements.txt
amqp==5.1.1
appdirs==1.4.4
argh==0.26.2
asgiref==3.6.0
attrs==22.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==21.6b0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.1.0
cryptography==39.0.0
defusedxml==0.7.1
Django==3.2.7
django-appconf==1.0.5
django-autoslug==1.9.8
django-celery-beat==2.4.0
django-celery-email==3.0.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-countries==7.2.1
django-environ==0.7.0
django-filter==21.1
django-phonenumber-field==5.2.0
django-templated-mail==1.1.1
django-timezone-field==5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==8.9.1
flake8==3.9.2
flower==0.9.7
humanize==4.5.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
isort==5.9.3
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.0
phonenumbers==8.12.33
Pillow==8.3.2
pluggy==1.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
PyJWT==2.1.0
pytest==7.2.1
pytest-cov==2.12.1
pytest-django==4.4.0
pytest-factoryboy==2.1.0
python-crontab==2.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==5.3.1
redis==3.5.3
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tornado==6.2
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
watchdog==1.0.2
wcwidth==0.2.6
